LPD-102667 Render the editor inline and let the host provide the modal
The editor no longer wraps itself in a ClayModal: ImageEditor is an inline component that fills its container, so the host decides whether it lives in a dialog or in a page. The bottom bar becomes a plain footer instead of a ClayModal.Footer, the styles hang from .image-editor, and the save controller calls onClose instead of a modal closer.

The sample opens the editor in a full-screen Clay modal of its own and keeps only what it needs: the portlet drops the content.Language bundle the module never ships, since every key lives in the global Language.properties and the title comes from the display name, and its save adapter turns the edited blob into a download without the ?save=slow and ?save=fail switches meant for an end-to-end suite that is not in the branch yet.
