+++ modules/apps/frontend-js/frontend-js-image-editor-web/src/main/resources/META-INF/resources/css/Stage.scss
@@ -2,7 +2,7 @@
  * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
-.image-editor-modal .editor-workspace {
+.image-editor .editor-workspace {
 	background: var(--editor-workspace-bg);
 	display: flex;
 	flex: 1;
@@ -10,13 +10,13 @@
 	scrollbar-gutter: stable;
 }
 
-.image-editor-modal .editor-workspace:focus-visible {
+.image-editor .editor-workspace:focus-visible {
 	box-shadow: inset 0 0 0 3px rgb(255 255 255 / 60%);
 	outline: 2px solid var(--editor-accent);
 	outline-offset: -2px;
 }
 
-.image-editor-modal .editor-stage {
+.image-editor .editor-stage {
 	box-sizing: content-box;
 	display: block;
 	margin: auto;
